Added script 'ip_conflict.py'
5 functions created:

1) unix_time_to_str(timestamp)
The function converts the Linux-epoch time to a easy readable string format

2) ip4_addresses()
The method returns a dictionary with adapter names as a key and a list of IP addresses as values

3) ip_duplication(ip, interface)
Linux only, package required: iputils-arping
The function pings the adapter address passed to it.
If there is a machine with the same address on the local network, it returns True, otherwise it returns False.

4) ip_duplicates()
The function checks all addresses of all adapters of this machine for collisions in the local network.
Returns a dictionary with keys as an adapter name, containing a dictionary with keys as IP addresses, the values of which are True or False

5) ip_duplicates_multiprocessing()
Same as 'ip_duplicates' function, but every address ping is a subprocess.